Moved basecoin server commands out of examples dir to share them
