--- examples/basecoin/cmd/basecoin/commands/ibc.go
@@ -1,8 +1,0 @@
-package commands
-
-// import "github.com/cosmos/cosmos-sdk/plugins/ibc"
-
-// // returns a new IBC plugin to be registered with Basecoin
-// func NewIBCPlugin() *ibc.IBCPlugin {
-// 	return ibc.New()
-// }
