+++ server/commands/init.go
