+++ server/commands/key.go
