+++ server/commands/reset.go
