+++ server/commands/root.go
